Fix minor areas where runtime platform was not agnostic regarding integration datatype
- by intention, the platform can use any data type as ‘number’, the type used to perform the integration. This is known to compile correctly for float, double and long double

- however, there are some places where higher-precision types cause compilation failures, apparently because the code implies a narrowing cast that seems to be permitted for long double but which is prevented for eg. boost::multiprecision types

- this patch contains fixes for some of these, but there are still errors in the integrator which mean that boost::multiprecision types do not yet compile

+++ CppTransport/templates/nontrivial_metric_core.h
@@ -1879,7 +1879,10 @@ namespace transport
 				// so something has gone wrong
 				if(iter == boost::end(range))
 					{
-            throw failed_to_compute_horizon_exit(tk->get_N_initial(), N_range, found_end, log_aH.size(), (N.size() > 0 ? N.back() : 0.0), (log_aH.size() > 0 ? log_aH.back() : 0.0), largest_k);
+            throw failed_to_compute_horizon_exit(tk->get_N_initial(), N_range, found_end, log_aH.size(),
+                                                 (N.size() > 0 ? N.back() : 0.0),
+                                                 (log_aH.size() > 0 ? static_cast<double>(log_aH.back()) : 0.0),
+                                                 largest_k);
 					}
 
         system.close_down_workspace();
